refactor: extract SystemFacade + ConfigFacade from CLIFacade
Split CLIFacade+Service.swift into two standalone facades:
- SystemFacade: service lifecycle, status, installer operations
- ConfigFacade: configuration, apply pipeline, TCP operations

CLIFacade now only hosts pack types and CLIVersion. All methods have
been extracted to domain-specific facades.

Closes #442
