+++ Sources/KeyPathCLI/Commands/Layer/LayerListCommand.swift
@@ -12,7 +12,7 @@ struct LayerList: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = ConfigFacade()
 
         let layers: [String]
         do {
