+++ Sources/KeyPathCLI/Commands/Service/ServiceRestartCommand.swift
@@ -12,7 +12,7 @@ struct ServiceRestart: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = SystemFacade()
 
         CLIOutput.progress("Restarting Kanata service...", context: ctx)
         let success = await facade.restartService()
