+++ Sources/KeyPathCLI/Commands/Service/ServiceStartCommand.swift
@@ -12,7 +12,7 @@ struct ServiceStart: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = SystemFacade()
 
         let success = await facade.startService()
         if success {
