+++ Sources/KeyPathCLI/Commands/System/SystemInspectCommand.swift
@@ -12,7 +12,7 @@ struct SystemInspect: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = SystemFacade()
         let result: CLIInspectResult
         do {
             result = try await withThrowingTimeout(seconds: globals.timeout) {
