+++ Sources/KeyPathCLI/Commands/System/SystemInstallCommand.swift
@@ -15,7 +15,7 @@ struct SystemInstall: AsyncParsableCommand {
         let spinner = CLISpinner(context: ctx)
         spinner.start("Installing...")
 
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = SystemFacade()
         let report: CLIInstallerReport
         do {
             report = try await withThrowingTimeout(seconds: globals.timeout) {
